refactor(wta): extract main.rs cli_tests into cli_tests.rs
Move the inline #[cfg(test)] mod cli_tests block out of main.rs into a
dedicated src/cli_tests.rs file declared via #[cfg(test)] mod cli_tests;.
Pure mechanical split; no behavior change. All 9 tests pass.
